pyproject.toml
[project]
name = "latex-email-daemon"
version = "0.1.0"
description = ""
authors = [
    {name = "Sermak",email = "egidius.mysliwietz@gmx.de"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imapclient (>=3.1.0,<4.0.0)",
    "pyzmail36 (>=1.0.5,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "latex_email_daemon", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
